+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # NEXT RELEASE
+* Add refreshMutedAccounts() method
+* **(BREAKING CHANGE)** Remove TheAlgorithm's userData property; add a getUserData() method that defers to MastoApi
 
 ### v0.40.8
 * (Demo App) link to repo in header